docker_compose/movies_admin/requirements.txt
django==4.0.4
uwsgi==2.0.20
python-dotenv==0.21.1
django-split-settings==1.2.0
psycopg2-binary==2.9.3
uWSGI==2.0.20
djangorestframework==3.13.1
markdown==3.4
django-filter==21.1
django-debug-toolbar==3.8.1
